backend/src/requirements.txt
pip-audit
python-binance
psycopg2-binary
bcrypt==4.2.1
cryptography
fastapi
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pytest
httpx
pytest-asyncio
aiofiles
pydantic-settings
redis
alembic
#ta-lib
pandas_ta
numpy==2.1.3
apscheduler
setuptools
isort
black
pydantic
